==> code/script4_individualtrajectories_DML.Rmd <==
---
title: "Individual Trajectories"
author: "Daniela Munoz Lopez"
date: "2024-10-25"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
#clear environment (RUN EACH TIME SCRIPT IS OPENED)
rm(list = ls())
```

#Load Packages
```{r}
#Load in packages that will be used throughout the markdown.
packages <- c("tidyverse",
                "summarytools",
                "psych",
                "reader",
                "tidyr",
                "lme4",
                "lmerTest",
                "jtools",
                "interactions",
                "rio",
                "ggplot2",
                "dplyr",
                "here",
                "emmeans",
                "sjPlot")
#invisible(lapply(packages, install.packages, character.only = TRUE)) #install packages (only run once)
invisible(lapply(packages, library, character.only = TRUE)) #load the packages
```

#Load data
```{r}
#load data.
data_dir = here() #Move to the digitalstress_p5_p6 folder (project folder)

data_folder <- "data" #Name of the folder containing the preprocessed data

data_folder_path <- file.path(data_dir, data_folder) #Path to the data within data folder

file_name <- "p5_p6_dss_promis_fully_scored.csv" #Name of the preprocessed data that was created in script1_datacleaning_DML

peru5_6_data <- read.csv(here(data_folder_path,file_name), header = T, sep = ",", na.strings=c("", " ","NA")) #Load data

nrow(peru5_6_data) #N = 973 -- ok: does match N in script 1

#check data
head(peru5_6_data, 20)
tail(peru5_6_data, 20)
```

#Recode wave to timepoint 
```{r}
peru5_6_data <- peru5_6_data %>% 
     dplyr::mutate_at(c('wave'), 
               funs(dplyr::recode(., "Peru 5"='Time 1',
                                     "Peru 6"='Time 2')))
```

#DEPRESSION
#Avail Stress
```{r}
#Model - avail stress
avail_stress_gen_grade_lm <- lm(avail_stress ~ wave + dep_clinical_status_invariant_p5, peru5_6_data)
summary(avail_stress_gen_grade_lm)
#Plot Availability Stress
p_avail <- plot_model(avail_stress_gen_grade_lm,
                type = "pred", terms = c("wave", "dep_clinical_status_invariant_p5")) + 
                ylab("Availability Stress") +
                xlab("Time Point") + labs(colour = "Depression Ranking") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(avail_stress_gen_grade_lm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, dep_clinical_status_invariant_p5, wave, avail_stress)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_avail_stress_gen_grade_lm <- predict(avail_stress_gen_grade_lm)


p_indiv_avail <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = avail_stress, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  theme_classic() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  labs(colour = "Depression Ranking") +
  ylab("Availability Stress") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_avail

#plot with averages
p_indiv_avail <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = avail_stress, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Availability Stress") + 
  labs(colour = "Depression Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_avail
```
#Approval Anxiety
```{r}
#Model - avail stress
approv_anx_gen_grade_lm_confirm <- lm(approval_anx ~ wave + dep_clinical_status_invariant_p5, peru5_6_data)
summary(approv_anx_gen_grade_lm_confirm)
#Plot Availability Stress
p_avail <- plot_model(approv_anx_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "dep_clinical_status_invariant_p5")) + 
                ylab("Approvavl Anxiety") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(approv_anx_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, dep_clinical_status_invariant_p5, wave, approval_anx)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_approv_anx_gen_grade_lm_confirm <- predict(approv_anx_gen_grade_lm_confirm)


p_indiv_anx <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = approval_anx, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Approval Anxiety") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_anx

#plot with averages
p_indiv_anx <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = approval_anx, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Approval Anxiety") + 
  labs(colour = "Depression Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_anx
```
#FOMO
```{r}
#Model - avail stress
fomo_gen_grade_lm_confirm <- lm(fomo ~ wave + dep_clinical_status_invariant_p5, peru5_6_data)
summary(fomo_gen_grade_lm_confirm)
#Plot Availability Stress
p_avail <- plot_model(fomo_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "dep_clinical_status_invariant_p5")) + 
                ylab("FOMO") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(fomo_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, dep_clinical_status_invariant_p5, wave, fomo)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_fomo_gen_grade_lm_confirm <- predict(fomo_gen_grade_lm_confirm)


p_indiv_fomo <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = fomo, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  theme_classic() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("FOMO") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_fomo

#plot with averages
p_indiv_fomo <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = fomo, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("FOMO") + 
  labs(colour = "Depression Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_fomo
```
#Online Vigilance
```{r}
#Model - avail stress
online_vigil_gen_grade_lm_confirm <- lm(online_vigil ~ wave + dep_clinical_status_invariant_p5, peru5_6_data)
summary(online_vigil_gen_grade_lm_confirm)
#Plot Availability Stress
p_avail <- plot_model(online_vigil_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "dep_clinical_status_invariant_p5")) + 
                ylab("Online Vigilance") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(online_vigil_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, dep_clinical_status_invariant_p5, wave, online_vigil)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_online_vigil_gen_grade_lm_confirm <- predict(online_vigil_gen_grade_lm_confirm)


p_indiv_ov <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = online_vigil, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Online Vigilance") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_ov

#plot with averages
p_indiv_ov <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = online_vigil, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Online Vigilance") + 
  labs(colour = "Depression Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_ov
```
#Connection Overload
```{r}
#Model - avail stress
connect_over_gen_grade_lm_confirm <- lm(connect_overload ~ wave + dep_clinical_status_invariant_p5, peru5_6_data)
summary(connect_over_gen_grade_lm_confirm)
#Plot Availability Stress
p_avail <- plot_model(connect_over_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "dep_clinical_status_invariant_p5")) + 
                ylab("Connection Overload") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(connect_over_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, dep_clinical_status_invariant_p5, wave, connect_overload)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_connect_over_gen_grade_lm_confirm <- predict(connect_over_gen_grade_lm_confirm)


p_indiv_co <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = connect_overload, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Connection Overload") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_co

#plot with averages
p_indiv_co <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = connect_overload, group = Peru3_4_5_6_ID, color=dep_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = dep_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Connection Overload") + 
  labs(colour = "Depression Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_co
```

#merge plots into one!!
```{r}
library(gridExtra)

dss_subscale_anx_dep <- grid.arrange(
  arrangeGrob(p_indiv_avail, p_indiv_anx, p_indiv_ov, ncol = 3),
  arrangeGrob(p_indiv_fomo, p_indiv_co, ncol = 3),
  nrow = 2
)

file.name = (here("output", "plots", "dss_subscale_anx_dep.jpeg"))
ggsave(file.name, plot = dss_subscale_anx_dep, dpi = 300, width = 14, height = 10, device = "jpeg")
```


#ANXIETY 
#Avail Stress
```{r}
#Model - avail stress
avail_stress_gen_grade_lm_confirm <- lm(avail_stress ~ wave + anx_clinical_status_invariant_p5, peru5_6_data)

#Plot Availability Stress
p_avail <- plot_model(avail_stress_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "anx_clinical_status_invariant_p5")) + 
                ylab("Availability Stress") +
                xlab("Time Point") + labs(colour = "Anxiety Ranking") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(avail_stress_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, anx_clinical_status_invariant_p5, wave, avail_stress)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_avail_stress_gen_grade_lm_confirm <- predict(avail_stress_gen_grade_lm_confirm)


p_indiv_avail <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = avail_stress, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  theme_classic() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  labs(colour = "Depression Ranking") +
  ylab("Availability Stress") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_avail

#plot with averages
p_indiv_avail <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = avail_stress, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Availability Stress") + 
  labs(colour = "Anxiety Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_avail
```
#Approval Anxiety
```{r}
#Model - avail stress
approv_anx_gen_grade_lm_confirm <- lm(approval_anx ~ wave + anx_clinical_status_invariant_p5, peru5_6_data)

#Plot Availability Stress
p_avail <- plot_model(approv_anx_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "anx_clinical_status_invariant_p5")) + 
                ylab("Approvavl Anxiety") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(approv_anx_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, anx_clinical_status_invariant_p5, wave, approval_anx)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_approv_anx_gen_grade_lm_confirm <- predict(approv_anx_gen_grade_lm_confirm)


p_indiv_anx <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = approval_anx, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Approval Anxiety") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_anx

#plot with averages
p_indiv_anx <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = approval_anx, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Approval Anxiety") + 
  labs(colour = "Anxiety Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_anx
```
#FOMO
```{r}
#Model - avail stress
fomo_gen_grade_lm_confirm <- lm(fomo ~ wave + anx_clinical_status_invariant_p5, peru5_6_data)

#Plot Availability Stress
p_avail <- plot_model(fomo_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "anx_clinical_status_invariant_p5")) + 
                ylab("FOMO") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(fomo_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, anx_clinical_status_invariant_p5, wave, fomo)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_fomo_gen_grade_lm_confirm <- predict(fomo_gen_grade_lm_confirm)


p_indiv_fomo <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = fomo, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  theme_classic() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("FOMO") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_fomo

#plot with averages
p_indiv_fomo <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = fomo, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("FOMO") + 
  labs(colour = "Anxiety Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_fomo
```
#Online Vigilance
```{r}
#Model - avail stress
online_vigil_gen_grade_lm_confirm <- lm(online_vigil ~ wave + anx_clinical_status_invariant_p5, peru5_6_data)

#Plot Availability Stress
p_avail <- plot_model(online_vigil_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "anx_clinical_status_invariant_p5")) + 
                ylab("Online Vigilance") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(online_vigil_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, anx_clinical_status_invariant_p5, wave, online_vigil)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_online_vigil_gen_grade_lm_confirm <- predict(online_vigil_gen_grade_lm_confirm)


p_indiv_ov <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = online_vigil, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Online Vigilance") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_ov

#plot with averages
p_indiv_ov <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = online_vigil, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Online Vigilance") + 
  labs(colour = "Anxiety Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_ov
```
#Connection Overload
```{r}
#Model - avail stress
connect_over_gen_grade_lm_confirm <- lm(connect_overload ~ wave + anx_clinical_status_invariant_p5, peru5_6_data)

#Plot Availability Stress
p_avail <- plot_model(connect_over_gen_grade_lm_confirm,
                type = "pred", terms = c("wave", "anx_clinical_status_invariant_p5")) + 
                ylab("Connection Overload") +
                xlab("Time Point") + labs(colour = "Sex") + theme(title= element_text(size=20, vjust=2, face="bold"),
                plot.title = element_blank(),
                axis.title.x= element_text(size=18, vjust=-0.3),
                axis.title.y= element_text(size=18, vjust=1.5),
                axis.text.x= element_text(size=16, colour="black"),
                axis.text.y= element_text(size=18, colour="black"),
                legend.title=element_text(size=16, colour = "black"),
                legend.position = "right",
                legend.text=element_text(size=16),
                strip.text = element_text(size=18, face="bold"),
                panel.background = element_blank(),
                axis.line = element_line(colour = "black"))  + geom_line() + 
                scale_y_continuous(breaks = seq(1, 5, 1),
                     limits = c(1, 5))
p_avail


#summary of winning model
summary(connect_over_gen_grade_lm_confirm)

#remove missing obs from vars in model
avail_vars <- peru5_6_data %>% select(Peru3_4_5_6_ID, anx_clinical_status_invariant_p5, wave, connect_overload)
avail_vars <- avail_vars[complete.cases(avail_vars), ] #1439

avail_vars$pred_connect_over_gen_grade_lm_confirm <- predict(connect_over_gen_grade_lm_confirm)


p_indiv_co <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = connect_overload, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.7, size = 0.5) + 
  
  theme_classic() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun = mean, geom = 'line', size = 0.5) + 
  xlab("Time Point") + 
  ylab("Connection Overload") + 
  theme(axis.text.x = element_text(size = 10), axis.text.y = element_text(size = 10), 
        axis.title=element_text(size=16, face = "bold"),legend.title = element_blank(), 
        legend.text=element_text(size=7), strip.text.x = element_text(size = 7.5)) + 
  #scale_color_manual(values=c("#D41159", "#1A85FF")) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5))
p_indiv_co

#plot with averages
p_indiv_co <- avail_vars %>% 
  ggplot(data = ., aes(x = wave, y = connect_overload, group = Peru3_4_5_6_ID, color=anx_clinical_status_invariant_p5)) + 
  geom_line(alpha = 0.1) + 
  theme_bw() + 
  stat_summary(aes(group = anx_clinical_status_invariant_p5), fun.y = mean, geom = 'line', size = 2) + 
  xlab("Time Point") + 
  ylab("Connection Overload") + 
  labs(colour = "Anxiety Ranking") + 
  theme(axis.text.x = element_text(size = 16), 
        axis.text.y = element_text(size = 16), 
        axis.title=element_text(size=16, face = "bold"), 
        legend.text=element_text(size=10), 
        strip.text.x = element_text(size = 7.5)) + 
  scale_y_continuous(breaks = seq(1, 5, 1),limits = c(1, 5)) + 
  scale_color_manual(values = c("clinical" = "#de0000",   # red
                                "non-clinical" = "#008000"),  # green
                     breaks = c("clinical", "non-clinical"))
p_indiv_co
```

#merge plots into one!!
```{r}
library(gridExtra)

ss_subscale_anx_only <- grid.arrange(
  arrangeGrob(p_indiv_avail, p_indiv_anx, p_indiv_ov, ncol = 3),
  arrangeGrob(p_indiv_fomo, p_indiv_co, ncol = 3),
  nrow = 2
)

file.name = (here("output", "plots", "dss_subscale_anx_only.jpeg"))
ggsave(file.name, plot = ss_subscale_anx_only, dpi = 300, width = 14, height = 10, device = "jpeg")
```
